+++ chia/util/timing.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 import os
 import sys
-from typing import Optional, overload
+import time
+from typing import Callable, Iterator, Optional, overload
 
 system_delays = {
     # based on data from https://github.com/Chia-Network/chia-blockchain/pull/13724
@@ -22,7 +23,7 @@ system_delays = {
 
 if os.environ.get("GITHUB_ACTIONS") == "true":
     # https://docs.github.com/en/actions/learn-github-actions/environment-variables#default-environment-variables
-    _system_delay = system_delays["github"][sys.platform]
+    pass
 else:
     _system_delay = system_delays["local"][sys.platform]
 
@@ -42,3 +43,21 @@ def adjusted_timeout(timeout: Optional[float]) -> Optional[float]:
         return None
 
     return timeout + _system_delay
+
+
+def backoff_times(
+    initial: float = 0.001,
+    final: float = 0.100,
+    time_to_final: float = 0.5,
+    clock: Callable[[], float] = time.monotonic,
+) -> Iterator[float]:
+    # initially implemented as a simple linear backoff
+
+    start = clock()
+    delta: float = 0
+
+    result_range = final - initial
+
+    while True:
+        yield min(final, initial + ((delta / time_to_final) * result_range))
+        delta = clock() - start
