+++ tests/core/util/test_file_keyring_synchronization.py
@@ -8,9 +8,9 @@ from pathlib import Path
 from sys import platform
 from time import sleep
 
-from chia.simulator.adjusted_timeout import adjusted_timeout
 from chia.simulator.keyring import TempKeyring
 from chia.util.keyring_wrapper import KeyringWrapper
+from chia.util.timing import adjusted_timeout
 from tests.core.util.test_lockfile import wait_for_enough_files_in_directory
 
 log = logging.getLogger(__name__)
