+++ tests/util/test_priority_mutex.py
@@ -13,8 +13,8 @@ from typing import Callable, List, Optional
 import anyio
 import pytest
 
-from chia.simulator.adjusted_timeout import adjusted_timeout
 from chia.util.priority_mutex import NestedLockUnsupportedError, PriorityMutex
+from chia.util.timing import adjusted_timeout
 from tests.util.misc import Marks, datacases
 from tests.util.time_out_assert import time_out_assert_custom_interval
